move initialize db from run to db, change imports, global variables
